[PLAT-13995][PLAT-12270][PLAT-14002][PLAT-14011] Tags should be optional when creating telemetry provider
Summary:
This diff fixes the following 4 tickets:
1. [PLAT-13995] Tags should be optional when creating telemetry provider.
2. [PLAT-12270] Rename logRow to logRows + pgaudit.log_row to pgaudit.log_rows in YBA code
3. [PLAT-14002] Should not allow deleting a Telemetry Provider when it's in use
4. [PLAT-14011] Add AuditLog Payload in ModifyAuditLogging task type details

For #2, it is okay to change the request body param `logRow` -> `logRows` since this is an internal API and no one uses this DB audit logs feature yet including UI, YBM, other clients, etc.

Test Plan:
Manually tested all 4 of the above test cases.
Run UTs.
Run itests.

Reviewers: amalyshev

Reviewed By: amalyshev

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35124

+++ managed/src/main/java/com/yugabyte/yw/controllers/TelemetryProviderController.java
@@ -28,12 +28,14 @@ import io.swagger.annotations.Authorization;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
+import lombok.extern.slf4j.Slf4j;
 import play.mvc.Http;
 import play.mvc.Result;
 
 @Api(
     value = "Telemetry Provider",
     authorizations = @Authorization(AbstractPlatformController.API_KEY_AUTH))
+@Slf4j
 public class TelemetryProviderController extends AuthenticatedController {
 
   @Inject private TelemetryProviderService telemetryProviderService;
@@ -126,7 +128,28 @@ public class TelemetryProviderController extends AuthenticatedController {
   })
   public Result deleteTelemetryProvider(
       UUID customerUUID, UUID providerUUID, Http.Request request) {
-    Customer.getOrBadRequest(customerUUID);
+    Customer customer = Customer.getOrBadRequest(customerUUID);
+
+    // Check if telemetry provider exists.
+    boolean doesTelemetryProviderExist =
+        telemetryProviderService.checkIfExists(customer.getUuid(), providerUUID);
+    if (!doesTelemetryProviderExist) {
+      String errorMessage =
+          String.format("Telemetry Provider '%s' does not exist.", providerUUID.toString());
+      log.error(errorMessage);
+      throw new PlatformServiceException(BAD_REQUEST, errorMessage);
+    }
+
+    // Check if telemetry provider is in use.
+    boolean isProviderInUse = telemetryProviderService.isProviderInUse(customer, providerUUID);
+    if (isProviderInUse) {
+      String errorMessage =
+          String.format(
+              "Cannot delete Telemetry Provider '%s', as it is in use.", providerUUID.toString());
+      log.error(errorMessage);
+      throw new PlatformServiceException(BAD_REQUEST, errorMessage);
+    }
+
     telemetryProviderService.delete(providerUUID);
     auditService()
         .createAuditEntryWithReqBody(
